Make x and y coordinates double

+++ lib/src/model/Coordinates.dart
@@ -1,8 +1,8 @@
 part of dozergame;
 
 class Coordinates {
-  final int x;
-  int y;
+  final double x;
+  double y;
   // final int dx; // by now only for extension possibilities, e.g. image direction
 
   Coordinates(this.x, this.y){}
