+++ lib/src/model/Entity.dart
@@ -5,16 +5,16 @@ abstract class Entity {
   int id;
 
   /** The current movement of the entity for the next view update */
-  int dx = 0;
-  int dy = 5;
+  double dx = 0;
+  double dy = 5;
 
-  int x;
-  int y = 0;
+  double x;
+  double y = 0;
 
   int width;
   int height;
 
-  void move(int dx, int dy) {
+  void move(double dx, double dy) {
     this.dx = dx;
     this.dy = dy;
   }
